requirements.txt
azure-storage-blob==12.5.0
azure-core==1.8.2
mock==4.0.2
amqp==2.6.0
asgiref==3.2.10
attrs==19.3.0
billiard==3.6.3.0
bootstrap4==0.1.0
celery==4.4.2
certifi==2020.6.20
chardet==3.0.4
coverage==5.2
defusedxml==0.6.0
Django==3.0.5
django-crispy-forms==1.9.1
django-filter==2.2.0
djangorestframework==3.11.0
django-bootstrap-breadcrumbs==0.9.2
docopt==0.6.2
docutils==0.16
gunicorn==19.10.0
html5lib==1.1
hypothesis==5.20.2
idna==2.10
importlib-metadata==1.7.0
jedi==0.15.2
json-merge-patch==0.2
json5==0.9.5
jsonschema==3.2.0
kombu==4.6.11
lockfile==0.12.2
more-itertools==8.4.0
numpy==1.19.0
packaging==20.4
pandas==1.0.5
parso==0.7.0
pep8==1.7.1
pipdeptree==1.0.0
pipreqs==0.4.10
pluggy==0.13.1
psutil==5.7.2
psycopg2-binary==2.8.5
py==1.9.0
py4j==0.10.7
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.20
pycrypto==2.6.1
pyodbc==4.0.30
pyparsing==2.4.7
pyrsistent==0.16.0
pyspark==2.4.5
pytest==5.4.3
pytest-arraydiff==0.3
pytest-astropy==0.8.0
pytest-astropy-header==0.1.2
pytest-cov==2.10.0
pytest-doctestplus==0.7.0
pytest-filter-subpackage==0.1.1
pytest-openfiles==0.5.0
pytest-remotedata==0.3.2
python-daemon==2.1.2
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-jsonrpc-server==0.3.4
python-language-server==0.31.10
python3-openid==3.1.0
python-decouple==3.3
pytz==2020.1
pytzdata==2019.3
PyYAML==5.3.1
regex==2020.6.8
requests==2.24.0
six==1.15.0
selenium==3.141.0
sortedcontainers==2.2.2
SQLAlchemy==1.3.18
sqlparse==0.3.1
ujson==1.35
unittest-xml-reporting==3.0.2
urllib3==1.25.9
uuid==1.30
VerbalExpressions==0.0.2
vine==1.3.0
wcwidth==0.2.5
webencodings==0.5.1
yarg==0.1.9
zipp==3.1.0
